# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalAction": {
        "description": "Public lowercase approval actions.",
        "enum": [
          "approve",
          "edit",
          "reject"
        ],
        "title": "ApprovalAction",
        "type": "string"
      },
      "ApprovalDetailResponse": {
        "additionalProperties": false,
        "description": "Authorized approval detail needed to render a human decision form.",
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "editable_fields": {
            "items": {
              "type": "string"
            },
            "title": "Editable Fields",
            "type": "array"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "planning_version": {
            "title": "Planning Version",
            "type": "integer"
          },
          "proposed_arguments": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Proposed Arguments",
            "type": "object"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "resolution_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Action"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "resolved_arguments": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Arguments"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          },
          "tool_version": {
            "title": "Tool Version",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "task_id",
          "status",
          "step_id",
          "planning_version",
          "tool_name",
          "tool_version",
          "editable_fields",
          "proposed_arguments",
          "resolved_arguments",
          "reason",
          "resolution_action",
          "resolution_reason",
          "created_at",
          "expires_at",
          "resolved_at",
          "resolved_by"
        ],
        "title": "ApprovalDetailResponse",
        "type": "object"
      },
      "ApprovalResolutionRequest": {
        "additionalProperties": false,
        "description": "One strictly validated approval resolution request body.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ApprovalAction"
          },
          "edited_arguments": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Arguments"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "action"
        ],
        "title": "ApprovalResolutionRequest",
        "type": "object"
      },
      "ApprovalResolutionResponse": {
        "additionalProperties": false,
        "description": "Stable decision acceptance; execution continues asynchronously when approved.",
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "approval_status": {
            "title": "Approval Status",
            "type": "string"
          },
          "resolution_action": {
            "title": "Resolution Action",
            "type": "string"
          },
          "resolved_at": {
            "format": "date-time",
            "title": "Resolved At",
            "type": "string"
          },
          "resolved_by": {
            "title": "Resolved By",
            "type": "string"
          },
          "resume_status": {
            "title": "Resume Status",
            "type": "string"
          },
          "runtime_status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_status": {
            "title": "Task Status",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "approval_status",
          "resolution_action",
          "task_id",
          "task_status",
          "runtime_status",
          "status_url",
          "resolved_at",
          "resolved_by",
          "resume_status",
          "trace_id"
        ],
        "title": "ApprovalResolutionResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "NaturalLanguageTaskSubmission": {
        "additionalProperties": false,
        "description": "Public request: natural-language task plus tightening-only execution options.",
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Metadata",
            "type": "object"
          },
          "output_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskOutputFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "read_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read Only"
          },
          "require_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Require Approval"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "task": {
            "minLength": 1,
            "title": "Task",
            "type": "string"
          },
          "task_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "task"
        ],
        "title": "NaturalLanguageTaskSubmission",
        "type": "object"
      },
...
      "RuntimeStatus": {
        "description": "Execution-host status kept separate from the frozen business TaskStatus.",
        "enum": [
          "READY",
          "LEASED",
          "WAITING_RETRY",
          "SUSPENDED",
          "FINISHED"
        ],
        "title": "RuntimeStatus",
        "type": "string"
      },
      "TaskErrorResponse": {
        "additionalProperties": false,
        "description": "Uniform transport error returned before a workflow result exists.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message",
          "trace_id"
        ],
        "title": "TaskErrorResponse",
        "type": "object"
      },
...
      "TaskListResponse": {
        "additionalProperties": false,
        "description": "Bounded current-user task history ordered newest-first.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
      "TaskOutputFormat": {
        "description": "Public output choices mapped onto the frozen Artifact contract.",
        "enum": [
          "pdf",
          "json"
        ],
        "title": "TaskOutputFormat",
        "type": "string"
      },
      "TaskResponse": {
        "additionalProperties": false,
        "description": "Stable public task-management summary.",
        "properties": {
          "artifact_count": {
            "title": "Artifact Count",
            "type": "integer"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "pending_approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Approval Id"
          },
          "runtime_status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "step_count": {
            "title": "Step Count",
            "type": "integer"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_summary": {
            "title": "Task Summary",
            "type": "string"
          },
          "task_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "trace_id",
          "status",
          "runtime_status",
          "task_type",
          "created_at",
          "started_at",
          "completed_at",
          "cancelled_at",
          "current_step",
          "task_summary",
          "pending_approval_id",
          "step_count",
          "evidence_count",
          "artifact_count",
          "error_summary"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "TaskStatus": {
        "description": "Authoritative lifecycle states from the frozen task state machine.",
        "enum": [
          "CREATED",
          "UNDERSTANDING",
          "PLANNING",
          "EXECUTING",
          "WAITING_APPROVAL",
          "RETRYING",
          "REPLANNING",
          "VERIFYING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
      "TaskSubmissionResponse": {
        "additionalProperties": false,
        "description": "Future 202 Accepted response returned after Task and dispatch commit atomically.",
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "artifacts_url": {
            "minLength": 1,
            "title": "Artifacts Url",
            "type": "string"
          },
          "runtime_status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "status_url": {
            "minLength": 1,
            "title": "Status Url",
            "type": "string"
          },
          "task_id": {
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          },
          "task_status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "trace_id": {
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "trace_id",
          "task_status",
          "runtime_status",
          "accepted_at",
          "status_url",
          "artifacts_url"
        ],
        "title": "TaskSubmissionResponse",
        "type": "object"
      },
      "TaskType": {
        "description": "Versioned business task classifications admitted by governed contracts.",
        "enum": [
          "supplier_quality_analysis.v1",
          "accounts_payable_analysis.v1"
        ],
        "title": "TaskType",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Agentic Enterprise Knowledge Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tasks": {
      "get": {
        "description": "Return one bounded tenant- and owner-scoped task history page.",
        "operationId": "list_tasks",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "description": "Durably accept one Task without executing LangGraph in the API process.",
        "operationId": "create_task",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NaturalLanguageTaskSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Submit Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "description": "Return an authorized task summary through the application service.",
        "operationId": "get_task",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}/approvals/{approval_id}": {
      "get": {
        "description": "Return the complete proposed input only after tenant and approver-role checks.",
        "operationId": "get_approval_v1_tasks__task_id__approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Approval",
        "tags": [
          "approvals"
        ]
      },
      "post": {
        "description": "Persist a decision and optional resume dispatch without running the Graph inline.",
        "operationId": "resolve_approval_v1_tasks__task_id__approvals__approval_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Resolve Approval",
        "tags": [
          "approvals"
        ]
      }
    },
...
  }
}